feat: migrate OID4VCSdJwtIssuingEndpointTest test to new framework
Closes #49411
